Continue work on tokenizer
Tokenizer for CON blocks seems to be working. I may be getting tired
of the <exception> library as the API is strange.

I decided to take a stab at a BNF for the Spin language, and I'm
also identifying language features that I'd like to change. I've
started implementing the ability to switch between blocks.